Add files via upload
To capture from the webcam: we’ll capture frames from the webcam using OpenCV. We create a VideoCapture object and set it to capture frames from the default camera (0). It sets the resolution of the webcam to 640x480. We then loop through the frames and display them in a window until the user presses ‘q’ to exit.